@@ -6,7 +6,8 @@ final _sharedPrefsProvider = FutureProvider<SharedPreferences>((_) {
   return SharedPreferences.getInstance();
 });
 
-final themeProvider = StateNotifierProvider<ThemeProvider>((ref) {
+final themeProvider =
+    StateNotifierProvider<ThemeProvider, AsyncValue<ThemeType>>((ref) {
   return ThemeProvider(ref.read);
 });
 
@@ -20,7 +21,7 @@ class ThemeProvider extends StateNotifier<AsyncValue<ThemeType>> {
 
   static const _themeKey = "theme";
 
-  SharedPreferences _sharedPreferences;
+  late SharedPreferences _sharedPreferences;
 
   void _init() {
     state = AsyncData(stringToTheme(_sharedPreferences.getString(_themeKey)));
